scenarios all good, spartans work well and fairly balanced, introduces enemy shooter at level 7, enemy tank at level 10, spartan scenario at level 8

+++ Projectfiles/Character.h
@@ -51,7 +51,7 @@
 @property int bulletType;
 @property int fallingHealth;
 @property double speed;
-@property int fallSpeed;
+@property double fallSpeed;
 @property bool attacked;
 @property double endgameFallSpeed;
